finish dsa asy, sy, stopped at mgm asy

+++ CADCOP/src/Messages/MsgAlgorithmFactor.java
@@ -5,7 +5,7 @@ import AgentsAbstract.NodeId;
 public class MsgAlgorithmFactor extends MsgAlgorithm{
 	
     public MsgAlgorithmFactor(NodeId senderNode, NodeId recieverNode, Object context, 
-    		double timeStamp) {
+    		int timeStamp) {
 		super(senderNode, recieverNode, context, timeStamp);
 		
 	}
