+++ CADCOP/src/Messages/MsgAnyTimeDown.java
@@ -4,8 +4,8 @@ import AgentsAbstract.NodeId;
 
 public class MsgAnyTimeDown extends MsgAnyTime {
 
-	public MsgAnyTimeDown(NodeId sender, NodeId reciever, Permutation context, double timeStamp,
-			double timeCreated) {
+	public MsgAnyTimeDown(NodeId sender, NodeId reciever, Permutation context, int timeStamp,
+			int timeCreated) {
 		super(sender, reciever, context, timeStamp);
 	}
